Add support for istio ingress with new structure
Previously Istio was defined as a boolean to be enabled or not. This was limiting, since Istio has concept of how the request should be routed (via their Gateway resource, ref doc https://istio.io/latest/docs/reference/config/networking/gateway/). Thus gateways should be able to be specified.
This change is related to https://github.com/InseeFrLab/onyxia-api/pull/242