Move service account specification from root to both api and web
Service account was previously specified at root level. When using serviceAccount.create, this was OK, since api and web would then get their own SAs. When serviceAccount.create was false however, the same SA would be used for both api and web, which is very problematic of several reasons.

This commit fix this issue by moving the serviceAccount field from root to both api and web. This also make it clearer what happen under-the-hood.
The change is breaking, and major version is therefore bumped.